+++ CHANGES.md
@@ -3,6 +3,7 @@ dev
 
 - add beta repository for switches if there is a dev release
   in any of the compilers for that switch.
+- Demote Debian 9 to a Tier 2 now that Debian 10 is stable.
 
 v6.2.0 2019-06-25 Cambridge
 ---------------------------